requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image processing
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4

# AI/ML
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.206

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Utils
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0

